gcc support part 2
now builds and runs with GCC13 on mingw32 just fine

fixed enough warnings to get rid of -fpermissive

fixed FName static init order bug that by sheer luck does not trigger on MSVC builds

fixed a bunch of FString printf fuck ups, god knows how this works on MSVC

+++ Source/Engine/Inc/UnConn.h
@@ -171,7 +171,7 @@ public:
 		debug(Index<UNetConnection::MAX_CHANNELS);
 		while
 		(	++Index<UNetConnection::MAX_CHANNELS
-		&&	(!Conn->Channels[Index] || Conn->Channels[Index]->ChType!=T::ChannelType) );
+		&&	(!Conn->Channels[Index] || (INT)Conn->Channels[Index]->ChType!=(INT)T::ChannelType) );
 	}
 	operator UBOOL()
 	{
